Improved blending and adjusted mouse control
Removed requirement to press mouse while dragging to change intensity.
Added range mapping function to set range of intensities. Added some
other folder paths for additional testing. Changed default intensity to
more realistic default.

+++ src/ofApp.h
@@ -20,6 +20,6 @@ class ofApp : public ofBaseApp{
 		void gotMessage(ofMessage msg);
     
         vector<ofImage> images;
-    int intensity;
+        float intensity;
 		
 };
